WIP: Fix unit type compilation errors
- Revert Dollar explicit conversion back to implicit for compatibility
- Add Value property to Dollar for explicit access when needed
- Update test parameter types from Dollar to decimal for xUnit compatibility
- Fix test assertions to handle unit type conversions properly

+++ tests/AutoBot.Tests/Services/ProcessTradeExecutionTests.cs
@@ -591,13 +591,13 @@ public class ProcessTradeExecutionTests
             TradeState.Open);
 
         // Assert
-        Assert.Equal(quantity, trade.quantity);
-        Assert.Equal(entryPrice, trade.price);
+        Assert.Equal(quantity, (decimal)trade.quantity);
+        Assert.Equal(entryPrice, (decimal)trade.price);
         Assert.Equal(leverage, trade.leverage);
         Assert.Equal(side.ToString().ToLower(), trade.side);
         Assert.Equal(expectedMargin, trade.margin);
         Assert.Equal(expectedPL, trade.pl);
-        Assert.Equal(expectedLiquidation, trade.liquidation);
+        Assert.Equal(expectedLiquidation, (decimal)trade.liquidation);
         Assert.Equal(expectedOpeningFee, trade.opening_fee);
         Assert.Equal(expectedClosingFee, trade.closing_fee);
     }
